pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "markdown>=3.7",
]
